refactor: 将 UID 识别逻辑从 Genshin 类移至 BvStatus 类
重构 UID 识别功能的实现位置，从 Genshin.Uid() 方法移动到 Bv.Uid() 静态方法，以统一视觉识别相关功能的管理。原 Genshin.Uid() 方法现在仅作为包装调用。这提高了代码的组织性，使相关功能更集中。
