+++ BetterGenshinImpact/Core/Script/Dependence/Genshin.cs
@@ -52,35 +52,9 @@ public class Genshin
     /// 通过 OCR 识别当前角色的 UID
     /// </summary>
     /// <returns>UID 数字，如果识别失败则返回 0</returns>
-    public async Task<int> Uid()
+    public Task<int> Uid()
     {
-        try
-        {
-            var x = 1683;
-            var y = 1051;
-            var width = 234;
-            var height = 28;
-            var scaleTo1080PRatio = this.ScaleTo1080PRatio;
-            
-            x = (int)Math.Round(x * scaleTo1080PRatio);
-            y = (int)Math.Round(y * scaleTo1080PRatio);
-            width = (int)Math.Round(width * scaleTo1080PRatio);
-            height = (int)Math.Round(height * scaleTo1080PRatio);
-            
-            using var region = CaptureToRectArea();
-            var recognitionObjectOcr = RecognitionObject.Ocr(x, y, width, height);
-            var res = region.Find(recognitionObjectOcr);
-            if (res?.Text == null) return 0;
-            var matches = Regex.Matches(res.Text, @"\d+");
-            if (matches.Count == 0) return 0;
-            var numberStr = string.Join("", matches.Select(m => m.Value));
-            return int.TryParse(numberStr, out var uid) ? uid : 0;
-        }
-        catch (Exception e)
-        {
-            Logger.LogError(e, "OCR 识别 UID 异常");
-            return 0;
-        }
+        return Task.FromResult(Bv.Uid());
     }
     
     public Lazy<NavigationInstance> LazyNavigationInstance { get; } = new(() =>
